more intelligent destroy

+++ destroy_cluster.sh
@@ -18,13 +18,17 @@
 [ -z $CONFIG_SH_SOURCED ] && source config.sh
 [ -z $FUNCTIONS_SH_SOURCED ] && source functions.sh
 
-set -x
+[ $DEBUG -gt 0 ] && set -x
 
 MACHINES=`VBoxManage list vms | grep $NODE_PREFIX | awk '{ gsub (/"/,""); print $1}'`
+RUNNING_MASCHINES=`VBoxManage list runningvms | grep $NODE_PREFIX | awk '{ gsub (/"/,""); print $1}'`
 
 for I in $MACHINES 
 do
-  VBoxManage controlvm $I poweroff
-  sleep 2
+  if echo $RUNNING_MASCHINES| grep $I &> /dev/null
+  then 
+    VBoxManage controlvm $I poweroff
+    sleep 2
+  fi
   VBoxManage unregistervm $I --delete
 done
